Implemented a much cleaner, concurrent control between multiple Sender and Receiver Instances on the same machine.
Actual NIO file transfer unimplemented.
Unable to break out of reader.ready() in the send() method of the Message class when calling it on a p2p socket between clients after one client disconnected without warning. No keepalive exists between clients. In turn, it's not possible to close the thread.